Group ajax related methods in a single controller
